Extra farming time parameter
Added a parameter to allow you to specify "extra farming time". This will be enforced whether or not a bone is actually found, as a cheap way to assure that you don't go past a zone before you're ready (by setting an unreasonable number into the extra time) or simply when you feel a little behind on prestiges or resources and want to catch up a little.